Added to the dashboard UI for overlay additions. Publish-Share/edit/delete buttons.
Input element for naming the overlay.

Switch to overlay editor instance when clicking the 'edit' button.

Status-bar color coded trim at the top of the overlay cards indicating if the card has unsaved changes.

Cards cannot be deleted without confirmation.

Editor instance is contained inside of an iframe to load the javascript that runs the rendering engine when it is needed.